refactor interfaces, services, OEWHeader
