+++ src/App.tsx
@@ -1,8 +1,5 @@
-import React, { useEffect, useState } from 'react'
 import './App.css'
-import MapboxMap from './components/organisms/MapboxMap'
-import TopBanner from './components/molecules/TopBanner'
-import Logo from './assets/logo.png'
+import MapboxMap from './components/organisms/MapBoxMap/MapboxMap'
 
 function App() {
     return (
