+++ src/interfaces/IDayOption.tsx
@@ -1,0 +1,4 @@
+export interface IDayOption {
+    readonly value: number
+    readonly label: string
+}
