+++ src/interfaces/IRegionProperties.tsx
@@ -1,4 +1,4 @@
-export interface RegionProperties {
+export interface IRegionProperties {
     id: number
     jobs: number[]
     name: string // todo not up to date
