+++ src/interfaces/api/IRegionDatetime.tsx
@@ -1,0 +1,16 @@
+import { Polygon } from 'geojson'
+
+export interface IImageData {
+    image_id: number
+    timestamp: number
+    geometry: Polygon
+}
+
+export interface IAPIRegionDatetimes {
+    [timestamp: string]: ImageData[]
+}
+
+export interface IRegionDatetime {
+    timestamp: number
+    imageData: ImageData[]
+}
