+++ src/services/regionLayerService.tsx
@@ -1,0 +1,49 @@
+import { FeatureCollection, GeoJsonProperties, Point } from 'geojson'
+import mapboxgl from 'mapbox-gl'
+
+function addRegionPopup(map: mapboxgl.Map) {
+    const popup = new mapboxgl.Popup({
+        closeButton: false,
+        closeOnClick: false,
+    })
+
+    map.on('mouseenter', 'regions', (event) => {
+        map.getCanvas().style.cursor = 'pointer'
+
+        if (event.features![0].geometry.type === 'Point') {
+            const coordinates = event.features![0].geometry.coordinates.slice() ?? []
+            const name = event.features![0].properties?.name
+            const area = event.features![0].properties?.area_km2
+            const description = `<strong>${name}</strong><br>Size of the region: ${area}km<sup>2</sup>`
+
+            while (Math.abs(event.lngLat.lng - coordinates[0]) > 180) {
+                coordinates[0] += event.lngLat.lng > coordinates[0] ? 360 : -360
+            }
+            popup.setLngLat([coordinates[0], coordinates[1]]).setHTML(description).addTo(map)
+        }
+    })
+    map.on('mouseleave', 'regions', () => {
+        map.getCanvas().style.cursor = ''
+        popup.remove()
+    })
+}
+
+export function addRegionLayer(map: mapboxgl.Map, regions: FeatureCollection<Point, GeoJsonProperties>) {
+    map.addSource('regions', {
+        type: 'geojson',
+        data: regions,
+    })
+
+    map.addLayer({
+        id: 'regions',
+        type: 'circle',
+        source: 'regions',
+        paint: {
+            'circle-color': '#ff0000',
+            'circle-radius': 9,
+            'circle-stroke-width': 6,
+            'circle-stroke-color': '#660000',
+        },
+    })
+    addRegionPopup(map)
+}
